split qwen36 moe host tensor helpers
